changed handleLaunch()
If Relego is open and "Send to Relego" is called from browser, Relego stays in background to not interupt the user's workflow.
If Relego is not open, it gets startet and stays open.

Check for launchParams changed to NOT check for !launchParams, incompatible to future webOS releases.

Both options raising a dashboard notification (same as last code).